ids are now size_t

+++ include/rt_camera_s.h
@@ -6,7 +6,7 @@
 /*   By: hush <hush@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/30 16:29:54 by hush              #+#    #+#             */
-/*   Updated: 2020/06/02 01:25:11 by hush             ###   ########.fr       */
+/*   Updated: 2020/06/02 01:31:05 by hush             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 typedef struct		s_camera
 {
-	long 			id;
+	size_t 			id;
 	t_vec			pos;
 	t_vec			plane_pos;
 	t_vec			direction;
